renamed files, organized category dir, fixed imports
